+++ src/primitives/interval.rs
@@ -1,19 +1,19 @@
 #[derive(Debug, Clone, PartialEq)]
 pub struct Interval {
-    pub min: f64,
-    pub max: f64,
+    pub min: f32,
+    pub max: f32,
 }
 
 impl Interval {
     #[must_use]
     #[inline]
-    pub const fn new(min: f64, max: f64) -> Self {
+    pub const fn new(min: f32, max: f32) -> Self {
         Self { min, max }
     }
 
     #[must_use]
     #[inline]
-    pub const fn contains(&self, x: f64) -> bool {
+    pub const fn contains(&self, x: f32) -> bool {
         self.min <= x && x <= self.max
     }
 }
@@ -21,6 +21,6 @@ impl Interval {
 // helper initializer to make code look pretty
 #[must_use]
 #[inline]
-pub const fn interval(min: f64, max: f64) -> Interval {
+pub const fn interval(min: f32, max: f32) -> Interval {
     Interval::new(min, max)
 }
